fix(skippy): keep cache ABI stable for carrier update
